pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Robin.Wang", email = "3Jg$94sbQ@q1"},
]
dependencies = [
    "django>=4.2.6",
    "pymysql>=1.1.0",
    "django-cors-headers>=4.2.0",
    "pyjwt>=2.8.0",
    "django-redis>=5.4.0",
    "ronglian-sms-sdk>=1.0",
    "requests>=2.31.0",
    "celery>=5.3.4",
    "gevent>=23.9.1",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
start = "doppler run -- python dashop/manage.py runserver"
mm = "doppler run -- python dashop/manage.py makemigrations"
mt = "doppler run -- python dashop/manage.py migrate"

